Fixed minor bug in line 70 of jgraph-core/src/main/java/org/jgrapht/alg/linkprediction/PreferentialAttachmentLinkPrediction.java

+++ jgrapht-core/src/main/java/org/jgrapht/alg/linkprediction/PreferentialAttachmentLinkPrediction.java
@@ -67,7 +67,7 @@ public class PreferentialAttachmentLinkPrediction<V, E>
         int du = graph.outDegreeOf(u);
         int dv = graph.outDegreeOf(v);
 
-        return du * dv;
+        return (double) du * dv;
     }
 
 }
